The testSimpleIfStatement() test from the ControlFlowTests class is failed if behind сurly braces of "if () {}  " operator there is located more than 4 whitespaces (e.g. "if () {}      " - 6 whitespace).
It is because the method ifThenElseBlockContinues() from the AbstractParser class is trying to check 'e' 'l' 's' 'e' letters behind the if-operator when the position of the cursor equal with the end of expression string.
It throws java.lang.ArrayIndexOutOfBoundsException.

It can easy check just adding more than 3 whitespace in existing source code and run this test:
original method (test - ok):

  public void testSimpleIfStatement() {
    test("if (true) { System.out.println(\"test!\") }  \n");
  }

after adding 4 whitespace (test - failed)

  public void testSimpleIfStatement() {
    test("if (true) { System.out.println(\"test!\") }      \n");
  }

One of the solution there is the checking that the position of cursor less at 4 charachter than end of string together with the checking 'e' 'l' 's' 'e' letters.

This happened on our two projects, when  "if () {}" operator was located at the end of MVEL expression and this code was reformatted and added more than 9 spaces behind сurly braces of if-operator.